vdl: a somewhat hacky solution (using buildtags) to bootstrap vdl generation (#178)
Prior to this PR, the vdl code generated required existing vdl generated output to compile. This PR breaks that dependency by using conditional compilation for bootstrapping the compiler. The dependencies are the v23/vdlroot/vdltool and v23/vdl packages. The former uses vdl to define the vdl.config file structure (in vdl) and the latter to define WireError. In both build tags are used to provide 'empty' definitions for these and the golang code generation is similarly altered to use them. The changes are little awkward since there is no clean API, but fairly small.
