+++ v23/vdlroot/vdltool/config.vdl
@@ -3,7 +3,8 @@
 // license that can be found in the LICENSE file.
 
 // Package vdltool defines types used by the vdl tool itself, including the
-// format of vdl.config files.
+// format of vdl.config files. No packages other than the vdltool command
+// line tool and v.io/x/ref/lib/vdl/... packages should use it.
 package vdltool
 
 // Config specifies the configuration for the vdl tool.  This is typically
