+++ x/ref/lib/vdl/compile/builtin.go
@@ -11,7 +11,7 @@ import (
 
 var (
 	// BuiltInPackage and BuiltInFile are used to hold the built-ins.
-	BuiltInPackage = newPackage("", "_builtin", "_builtin", vdltool.Config{})
+	BuiltInPackage = newPackage("", "_builtin", "_builtin", vdltool.Config{}, "")
 	BuiltInFile    = &File{BaseName: "_builtin.vdl"}
 )
