+++ x/ref/lib/vdl/parse/grammar.y.go
@@ -9,14 +9,13 @@
 
 package parse
 
-import __yyfmt__ "fmt"
-
-//line grammar.y:20
-
 import (
+	__yyfmt__ "fmt"
 	"math/big"
 	"strings"
 )
+
+//line grammar.y:20
 
 type intPos struct {
 	int *big.Int
